confirmation de suppression d'une categorie

+++ src/app/admin/categories/categories.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog, MatDialogActions } from '@angular/material/dialog';
 import { AddEditCatComponent } from '../add-edit-cat/add-edit-cat.component';
 import { CategorieService } from 'src/app/services/categorie.service';
+import Swal from 'sweetalert2';
 
-@Component({ 
+@Component({
   selector: 'app-categories',
   templateUrl: './categories.component.html',
   styleUrls: ['./categories.component.css'],
@@ -43,26 +44,49 @@ export class CategoriesComponent implements OnInit {
     });
   }
   deleteCategorie(id: number) {
-    this.categorieService.deleteCategorie(id).subscribe({
-      next: (res) => {
-        alert('Categorie supprimé avec succès !');
-        this.getCategoriesList();
-      },
-      error: console.log,
+    Swal.fire({
+      title: 'Êtes-vous sûr?',
+      text: 'Cette Catégorie va être supprimé définitivement! ',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, supprimez-la!',
+      cancelButtonText: 'Annuler',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.categorieService.deleteCategorie(id).subscribe({
+          next: (res) => {
+            Swal.fire(
+              'Supprimé!',
+              'Votre Catégorie a été supprimé.',
+              'success'
+            );
+            this.getCategoriesList();
+          },
+          error: (err) => {
+            Swal.fire(
+              'Erreur!',
+              'Une erreur est survenue lors de la suppression.',
+              'error'
+            );
+          },
+        });
+      }
     });
   }
-  openEditCatForm( data : any ) {
+  openEditCatForm(data: any) {
     const dialogRef = this.dialog.open(AddEditCatComponent, {
       data,
       width: '550px',
     });
-      dialogRef.afterClosed().subscribe({
-        next: (val) => {
-          if (val) {
-            this.getCategoriesList();
-          }
-        },
-        error: console.log,
-      });
-  } 
+    dialogRef.afterClosed().subscribe({
+      next: (val) => {
+        if (val) {
+          this.getCategoriesList();
+        }
+      },
+      error: console.log,
+    });
+  }
 }
